MODFISTO-90 Support all-or-nothing operations for pending payments by invoice
- deduplicate
