+++ src/main/java/org/folio/service/transactions/PaymentCreditAllOrNothingService.java
@@ -42,7 +42,6 @@ import org.folio.service.fund.FundService;
 import org.folio.service.ledger.LedgerService;
 import org.folio.service.ledgerfy.LedgerFiscalYearService;
 import org.folio.service.summary.TransactionSummaryService;
-import org.javamoney.moneta.Money;
 
 import io.vertx.core.Context;
 import io.vertx.core.Future;
@@ -445,32 +444,6 @@ public class PaymentCreditAllOrNothingService extends BaseAllOrNothingTransactio
   @Override
   String getSummaryId(Transaction transaction) {
     return transaction.getSourceInvoiceId();
-  }
-
-  /**
-   * Calculates remaining amount for payment
-   * [remaining amount] = (allocated * allowableExpenditure) - (allocated - (unavailable + available)) - (awaitingPayment + expended + encumbered)
-   *
-   * @param budget     processed budget
-   * @param currency   processed transaction currency
-   * @return remaining amount for payment
-   */
-  @Override
-  protected Money getBudgetRemainingAmount(Budget budget, String currency) {
-    Money allocated = Money.of(budget.getAllocated(), currency);
-    // get allowableExpenditure from percentage value
-    double allowableExpenditure = Money.of(budget.getAllowableExpenditure(), currency).divide(100d).getNumber().doubleValue();
-    Money unavailable = Money.of(budget.getUnavailable(), currency);
-    Money available = Money.of(budget.getAvailable(), currency);
-    Money expenditure = Money.of(budget.getExpenditures(), currency);
-    Money awaitingPayment = Money.of(budget.getAwaitingPayment(), currency);
-    Money encumbered = Money.of(budget.getEncumbered(), currency);
-
-    Money result = allocated.multiply(allowableExpenditure);
-    result = result.subtract(allocated.subtract(unavailable.add(available)));
-    result = result.subtract(expenditure.add(awaitingPayment).add(encumbered));
-
-    return result;
   }
 
   @Override
